third_party/blink/web_tests/fast/frames/location-change-no-file-access-expected.txt

CONSOLE MESSAGE: line 4: SUCCESS
